Refs #7116. Separating RTPSMessageGroup_t into a private header.
