+++ src/cpp/rtps/messages/RTPSMessageGroup.cpp
@@ -23,6 +23,7 @@
 #include "../participant/RTPSParticipantImpl.h"
 #include "../flowcontrol/FlowController.h"
 #include "RTPSGapBuilder.hpp"
+#include "RTPSMessageGroup_t.hpp"
 
 #include <fastrtps/log/Log.h>
 
@@ -31,54 +32,6 @@
 namespace eprosima {
 namespace fastrtps {
 namespace rtps {
-
-/**
- * Class RTPSMessageGroup_t that contains the messages used to send multiples changes as one message.
- * @ingroup WRITER_MODULE
- */
-class RTPSMessageGroup_t
-{
-    public:
-
-        RTPSMessageGroup_t()
-            : rtpsmsg_submessage_(0u)
-            , rtpsmsg_fullmsg_(0u)
-#if HAVE_SECURITY
-            , rtpsmsg_encrypt_(0u)
-#endif
-        {
-        }
-
-        void init(
-#if HAVE_SECURITY
-                bool has_security,
-#endif
-                uint32_t payload,
-                const GuidPrefix_t& participant_guid)
-        {
-
-            rtpsmsg_fullmsg_.reserve(payload);
-            rtpsmsg_submessage_.reserve(payload);
-
-#if HAVE_SECURITY
-            if (has_security)
-            {
-                rtpsmsg_encrypt_.reserve(payload);
-            }
-#endif
-
-            CDRMessage::initCDRMsg(&rtpsmsg_fullmsg_);
-            RTPSMessageCreator::addHeader(&rtpsmsg_fullmsg_, participant_guid);
-        }
-
-        CDRMessage_t rtpsmsg_submessage_;
-
-        CDRMessage_t rtpsmsg_fullmsg_;
-
-#if HAVE_SECURITY
-        CDRMessage_t rtpsmsg_encrypt_;
-#endif
-};
 
 static thread_local std::unique_ptr<RTPSMessageGroup_t> tls_group;
